+++ SparkKit/Sources/Plugin API/SparkFunctions.h
@@ -50,18 +50,18 @@ void SparkDisplayAlert(SparkAlert *alert) {
 #pragma mark Notification
 /* delay: -1 for default value */
 SPARK_EXPORT
-void SparkNotificationDisplayIcon(IconRef icon, float delay);
+void SparkNotificationDisplayIcon(IconRef icon, CGFloat delay);
 
 SPARK_EXPORT
-void SparkNotificationDisplayImage(NSImage *anImage, float delay);
+void SparkNotificationDisplayImage(NSImage *anImage, CGFloat delay);
 
 /* See <HIServices/Icons.h> for possible values */
 SPARK_EXPORT
-void SparkNotificationDisplaySystemIcon(OSType icon, float delay);
+void SparkNotificationDisplaySystemIcon(OSType icon, CGFloat delay);
 
 /* Private notifications */
 SPARK_EXPORT
-void SparkNotificationDisplay(NSView *view, float delay);
+void SparkNotificationDisplay(NSView *view, CGFloat delay);
 
 #endif
